+++ R/get.R
@@ -369,6 +369,23 @@ getMetaFeaturesLinkouts <- function(study, modelID = NULL, quiet = FALSE, librar
   )
 }
 
+#' Get metaAssays from a study
+#'
+#' @inherit shared-get
+#' @inheritParams listStudies
+#'
+#' @export
+getMetaAssays <- function(study, modelID = NULL, quiet = FALSE, libraries = NULL) {
+  getElements(
+    study,
+    elements = "metaAssays",
+    filters = list(modelID = modelID),
+    default = "default",
+    quiet = quiet,
+    libraries = libraries
+  )
+}
+
 # ... Arguments passed to either data.table::fread() or jsonlite::read_json()
 getElements <- function(
   study,
@@ -459,13 +476,13 @@ getElements.character <- function(
   ...
 )
 {
-  oaDirectory <- getDirectory(study, libraries)
-  if (oaDirectory == "") {
+  onDirectory <- getDirectory(study, libraries)
+  if (onDirectory == "") {
     stop(sprintf("The study \"%s\" is not installed\n", study),
          "Did you run installStudy()?\n")
   }
 
-  elementsDirectory <- file.path(oaDirectory, elements)
+  elementsDirectory <- file.path(onDirectory, elements)
   if (!dir.exists(elementsDirectory)) {
     if (!quiet) message(sprintf("Study \"%s\" does not have any elements named \"%s\"",
                                 study, elements))